lss_cloudfunction/lss-admin-cf/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==1.9.0
flask
google-cloud-pubsub==2.1.0